Fix up f03ac96658
We need to cast the value returned by hexdec() to an Integer because
on 32bit PHP systems values higher than 2147483647 get converted
to Floats, which, when converted back to Integers, might become negative.
Therefore we cast the value to an Integer before abs() is being invoked on
it, else it would return the absolute value of a Float which might
in turn become negative after casting to an Integer (used as an array index).